product-market-research-crew/pyproject.toml
[project]
name = "product-market-research-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.108.0",
    "crewai-tools>=0.38.1",
    "groq>=0.20.0",
    "langtrace-python-sdk>=3.8.9",
    "python-dotenv>=1.1.0",
    "scrapegraph-py>=1.12.0",
    "tavily-python>=0.5.2",
]
